% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmean_util.R
\name{bic_kmean_est}
\alias{bic_kmean_est}
\title{Estimate a partially functional linear regression model with latent group structures.}
\usage{
bic_kmean_est(data_list, num_group = 2, num_pca = 5,
  est_fix_eff = TRUE, max_iter = 100, loc_search = FALSE,
  group_index = NULL)
}
\arguments{
\item{data_list}{A list of data. Several elements must be present in the list. The reponse \code{y},
the functional covariate \code{x_recv}, the scalar covariates \code{z}, and an index matrix \code{index}. 
The functional covariate \code{x_recv} must be generated from the \code{fda} package by, e.g., spline smoothing. 
The scalar covariates \code{z} is a matrix. The index matrix \code{index} is a data.frame recording the structure of the
data. The first column of \code{index} is the family number, the second column is the within family index. The column names of \code{index}
must be \code{ind_b} and \code{ind_w}.}

\item{num_group}{A vector of candidate number of groups.}

\item{num_pca}{A vector of candidate number of principal components.}

\item{est_fix_eff}{A logical value. If \code{TRUE}, then the fixed effects are estimated. Otherwise,
the fixed effects are not estimated}

\item{max_iter}{The maximum number of iteration. Default to be 100.}

\item{loc_search}{A logical value. If \code{TRUE}, conduct a local search to decrease the objective function.}

\item{group_index}{An initial value of the group membership, optional.}
}
\description{
This is the proposed estimator, which is developed based on the functional principal component analysis.
The algorithm is a variant of the K-means clustering algorithm. 
The number of groups and the number of principal components are determined by the BIC criterion.
}
