% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pena_utility.R
\name{family_wise}
\alias{family_wise}
\title{Estimate a partially functional linear regression model with family-specific functional coefficients.}
\usage{
family_wise(data_list, num_pca_vec, est_fix_eff = TRUE)
}
\arguments{
\item{data_list}{A list of data. Several elements must be present in the list. The reponse \code{y},
the functional covariate \code{x_recv}, the scalar covariates \code{z}, and an index matrix \code{index}. 
The functional covariate \code{x_recv} must be generated from the \code{fda} package by, e.g., spline smoothing. 
The scalar covariates \code{z} is a matrix. The index matrix \code{index} is a data.frame recording the structure of the
data. The first column of \code{index} is the family number, the second column is the within family index. The column names of \code{index}
must be \code{ind_b} and \code{ind_w}.}

\item{est_fix_eff}{A logical value. If \code{TRUE}, then the fixed effects are estimated. Otherwise,
the fixed effects are not estimated}

\item{num_pca}{A vector of candidate number of principal components.}
}
\description{
The estimator assumes family-specific functional coefficients.  
The the number of principal components are determined by the BIC criterion.
}
