% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pena_utility.R
\name{naive_kmeans}
\alias{naive_kmeans}
\title{Estimate a partially functional linear regression model with latent group structures using a two-step method.}
\usage{
naive_kmeans(data_list, num_group_vec, num_pca_vec, est_fix_eff = TRUE)
}
\arguments{
\item{data_list}{A list of data. Several elements must be present in the list. The reponse \code{y},
the functional covariate \code{x_recv}, the scalar covariates \code{z}, and an index matrix \code{index}. 
The functional covariate \code{x_recv} must be generated from the \code{fda} package by, e.g., spline smoothing. 
The scalar covariates \code{z} is a matrix. The index matrix \code{index} is a data.frame recording the structure of the
data. The first column of \code{index} is the family number, the second column is the within family index. The column names of \code{index}
must be \code{ind_b} and \code{ind_w}.}

\item{num_group_vec}{A vector of candidate number of groups.}

\item{num_pca_vec}{A vector of candidate number of principal components.}

\item{est_fix_eff}{A logical value. If \code{TRUE}, then the fixed effects are estimated. Otherwise,
the fixed effects are not estimated}
}
\description{
In the first step, a family-wise estimator is calculated. In the second step, the classic K-means algorithm
is applied to cluster the families.
The number of groups and the number of principal components are determined by the BIC criterion.
}
