% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pena_utility.R
\name{pena_est_fda_scale}
\alias{pena_est_fda_scale}
\title{Estimate a partially functional linear regression model with latent group structures using penalization methods.}
\usage{
pena_est_fda_scale(data_list, num_pca_vec, penalty = "SCAD",
  est_fix_eff = TRUE, precision = c(1e-05, 1e-05),
  lam_max_interval = c(0.1, 10), lam_len = 100, lamGroup_a = NULL,
  scale_pena = FALSE)
}
\arguments{
\item{data_list}{A list of data. Several elements must be present in the list. The reponse \code{y},
the functional covariate \code{x_recv}, the scalar covariates \code{z}, and an index matrix \code{index}. 
The functional covariate \code{x_recv} must be generated from the \code{fda} package by, e.g., spline smoothing. 
The scalar covariates \code{z} is a matrix. The index matrix \code{index} is a data.frame recording the structure of the
data. The first column of \code{index} is the family number, the second column is the within family index. The column names of \code{index}
must be \code{ind_b} and \code{ind_w}.}

\item{penalty}{A string indicates which penalty function to use, SCAD or LASSO.}

\item{est_fix_eff}{A logical value. If \code{TRUE}, then the fixed effects are estimated. Otherwise,
the fixed effects are not estimated}

\item{precision}{A vector of thresholds for terminating the ADMM algorithm.}

\item{lam_max_interval}{A vector indicating the search limits for the maximum of lambda such all the families are clutered into one group.}

\item{lam_len}{The size of grid for searching lambda.}

\item{lamGroup_a}{A vecotr of candidate values for lambda. If not provided, the algorithm will determine it automatically.}

\item{scale_pena}{A logical value, whether to scale the covariates before applying the ADMM algorithm.}

\item{num_pca}{A vector of candidate number of principal components.}
}
\description{
The estimator is a variant of the penalization estimator implemented with a ADMM algorithm. 
The number of groups and the number of principal components are determined by the BIC criterion.
}
